threatscope-ai/requirements.txt
 
fastapi
uvicorn
python-dotenv
google-generativeai
